added framework for saving data to a database
in the CS_SaveSystem.cs, I created a function called pushToDB. Feel free to change it as you see fit to get it to work with firebase. Right now it is a dummy function that is called automatically when game data is saved. Finally, the location for the saved game is based on Unity's built in "Application.persistentDataPath" string. This is so that the location of game data is consistent on any computer. My default location is in my users/.../AppData/LocalLow/DefaultCompany/WebWordle/saveData.json. I have a debug.log of the file path printed in unity console when ever a game instance is loaded (as in a game mode is selected)