added background image scrolling effect to main menu (#65)
* Pipeline Update

* Pipeline patch

* Patch for google credentials

* Typo

* Google credentials

* Patch

* Patch

* Patch

* Secret

* fixed save game

-created new, serializable script called gameModeData
-removed "[system.serializable]" from cs_mainmenu script.
-replaced all references to cs_mainmenu class object to the cs_gameModeData object for save and load functions

* New Firebase Pipeline

* Remove old firebase pipeline files

* Rename Private Key.json to PrivateKey.json

Rename file to be deleted.

* Remove Private Key

* -added easy and hard game modes

-need to fix a bug where the {timer} remains active when timer mode is disabled
-added gamemode loading functionality
- easy mode (arbitrary config: 4 letters, 6 tries)
- hardmode (arbitrary config: 7 letters, 4 tries)

* -removed extra build folders

* - created new build of project in codebusterweb/build folder

* Basic Achievements

Added a few basic achievements, tracked using boolean variables on the AchievementGet Object

* added save data system

* Fixes

- Fixed earning the same achievements multiple times
- Fixed checking for "Three birds" achievement (three new correct letters guessed in one guess)
- Fixed timer not working in classic mode

* added framework for saving data to a database

in the CS_SaveSystem.cs, I created a function called pushToDB. Feel free to change it as you see fit to get it to work with firebase. Right now it is a dummy function that is called automatically when game data is saved. Finally, the location for the saved game is based on Unity's built in "Application.persistentDataPath" string. This is so that the location of game data is consistent on any computer. My default location is in my users/.../AppData/LocalLow/DefaultCompany/WebWordle/saveData.json. I have a debug.log of the file path printed in unity console when ever a game instance is loaded (as in a game mode is selected)

* Update CS_SaveSystem.cs

* -removed merge conflict markers

* -removed shortcutmanagement library because it prevented web gl build

* Added sound effects and particles

- also fixed the three birds achievement for real this time (i hope)

* few sounds

* added scrolling background effect

created cs_scrolling.cs script with variables that can scale the scrolling rate

---------

Co-authored-by: girlReWired <63616817+girlReWired@users.noreply.github.com>
Co-authored-by: baid <cvscholtisek@alaska.edu>